@@ -4,22 +4,23 @@
   "author": "K3NN37H",
   "license": "GPL-3.0-only",
   "version": "0.1.1",
-  "packageManager": "yarn@3.3.0",
+  "type": "module",
+  "packageManager": "yarn@4.6.0",
   "scripts": {
     "start": "vite",
     "build": "vite build",
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@types/react": "^17.0.27",
-    "@types/react-dom": "^17.0.10",
-    "@vitejs/plugin-react": "^3.0.0",
-    "typescript": "^4.4.3",
-    "vite": "^4.0.1"
+    "@types/react": "^19.0.4",
+    "@types/react-dom": "^19.0.2",
+    "@vitejs/plugin-react": "^4.3.4",
+    "typescript": "^5.7.3",
+    "vite": "^6.0.7"
   },
   "dependencies": {
     "codepage": "^1.15.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   }
 }
